+++ SM_Il2Cpp/Main.cs
@@ -24,6 +24,12 @@ namespace MelonLoader.Support
         private static ISupportModule_To Initialize(ISupportModule_From interface_from)
         {
             Interface = interface_from;
+
+            MelonPreferences.AddEntryTypeResolveEvent(Preferences.Types.UnityEngine_Color.Resolve);
+            MelonPreferences.AddEntryTypeResolveEvent(Preferences.Types.UnityEngine_Vector2.Resolve);
+            MelonPreferences.AddEntryTypeResolveEvent(Preferences.Types.UnityEngine_Vector3.Resolve);
+            MelonPreferences.AddEntryTypeResolveEvent(Preferences.Types.UnityEngine_Vector4.Resolve);
+
             LogSupport.RemoveAllHandlers();
             if (MelonDebug.IsEnabled())
                 LogSupport.InfoHandler += MelonLogger.Msg;